problem100/p111_minimum_depth_of_binary_tree.go
package problem100

/**
111. 二叉树的最小深度
给定一个二叉树，找出其最小深度。
最小深度是从根节点到最近叶子节点的最短路径上的节点数量。
说明：叶子节点是指没有子节点的节点。
示例 1：
输入：root = [3,9,20,null,null,15,7]
输出：2

示例 2：
输入：root = [2,null,3,null,4,null,5,null,6]
输出：5
 * Definition for a binary tree node.
 * type TreeNode struct {
 *     Val int
 *     Left *TreeNode
 *     Right *TreeNode
 * }
*/
func minDepth(root *TreeNode) int {
	//关键在于理解：当前节点左右子节点都为空时才是叶子节点
	if root == nil {
		return 0
	}
	if root.Left == nil && root.Right == nil {
		return 1
	}

	if root.Left == nil {
		return 1 + minDepth(root.Right)
	}
	if root.Right == nil {
		return 1 + minDepth(root.Left)
	}

	depthLeft := minDepth(root.Left)
	depthRight := minDepth(root.Right)
	if depthLeft > depthRight {
		return 1 + depthRight
	} else {
		return 1 + depthLeft
	}
}
